+++ src/main/scala/geotrellis/data/png/Compression.scala
@@ -1,0 +1,4 @@
+package geotrellis.data.png
+
+sealed abstract class Compression(val n:Byte)
+case object Deflate extends Compression(0)
