+++ src/main/scala/geotrellis/data/png/Const.scala
@@ -1,0 +1,7 @@
+package geotrellis.data.png
+
+object Const {
+  // color modes for png
+  final val rgb = 2
+  final val rgba = 6
+}
