+++ src/main/scala/geotrellis/data/png/Settings.scala
@@ -1,0 +1,3 @@
+package geotrellis.data.png
+
+case class Settings(color:Color, filter:Filter)
